Add option to post populate IntendedFor fields
- Flywheel's BIDS Curation gear is currently somewhat limited and can
not handle multiple runs. This means if any manual curation of the BIDS
file paths occurs after BIDS Curation, the metadata for the IntendedFor
fields will be wrong. This commit adds an option to populate the
IntendedFor fields AFTER having downloaded your BIDS data set. This
ensures that any changes to the BIDS names are propagated to the fmap
sidecars without requiring manual editing of the metadata, which can be
very tedious.